deployer/requirements.txt
flask
requests
paramiko
pathlib
kafka-python